Moved activity to activity package
