+++ ShinyExpenseTracker/src/ca/ualberta/cs/shinyexpensetracker/activities/ClaimSummaryFragment.java
@@ -36,6 +36,5 @@ public class ClaimSummaryFragment extends Fragment {
 				container, false);
 		return rootView;
 	}
-}
 	
-
+}
